fix: remove potongan gaji relationship with jabatan

+++ app/Http/Controllers/PotonganGajiController.php
@@ -13,20 +13,10 @@ class PotonganGajiController extends Controller
      */
     public function index(Request $req)
     {
-        $jabatan_id = $req->query('jabatan_id');
-        $jabatan = Jabatan::find($jabatan_id);
-        $list_jabatan = Jabatan::all();
-
-        if($jabatan == null) {
-            return view('potongan_gaji.form', ['list_jabatan' => $list_jabatan]);
-        } else {
-            $list_potongan_gaji = PotonganGaji::where('jabatan_id', $jabatan->id)->get();
-            return view('potongan_gaji.index', [
-                'list_jabatan' => $list_jabatan,
-                'jabatan' => $jabatan,
-                'list_potongan_gaji' => $list_potongan_gaji,
-            ]);
-        }
+        $list_potongan_gaji = PotonganGaji::all();
+        return view('potongan_gaji.index', [
+            'list_potongan_gaji' => $list_potongan_gaji,
+        ]);
     }
 
     /**
@@ -34,9 +24,7 @@ class PotonganGajiController extends Controller
      */
     public function create()
     {
-        
-        $list_jabatan = Jabatan::all();
-        return view('potongan_gaji.create', ['list_jabatan' => $list_jabatan]);
+        return view('potongan_gaji.create');
     }
 
     /**
@@ -45,14 +33,12 @@ class PotonganGajiController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-            'jabatan_id' => 'required|exists:'.Jabatan::class.',id',
-            'keterangan' => 'required|string',
-            'nilai_potongan' => 'required|numeric'
+            'keterangan' => 'required|string|unique:'.PotonganGaji::class,
+            'nilai_potongan' => 'required|numeric|gte:0'
         ]);
 
         $pg = new PotonganGaji();
         $pg->keterangan = $validated['keterangan'];
-        $pg->jabatan_id = $validated['jabatan_id'];
         $pg->nilai_potongan = $validated['nilai_potongan'];
         $pg->save();
 
@@ -72,10 +58,7 @@ class PotonganGajiController extends Controller
      */
     public function edit(PotonganGaji $potonganGaji)
     {
-        
-        
-        $list_jabatan = Jabatan::all();
-        return view('potongan_gaji.create', ['list_jabatan' => $list_jabatan, 'potongan_gaji' => $potonganGaji]);
+        return view('potongan_gaji.edit', ['potongan_gaji' => $potonganGaji]);
     }
 
     /**
@@ -83,15 +66,24 @@ class PotonganGajiController extends Controller
      */
     public function update(Request $request, PotonganGaji $potonganGaji)
     {
-        
-        $validated = $request->validate([
-            'jabatan_id' => 'required|exists:'.Jabatan::class.',id',
-            'keterangan' => 'required|string',
-            'nilai_potongan' => 'required|numeric'
-        ]);
+        $validated = [];
 
-        $potonganGaji->keterangan = $validated['keterangan'];
-        $potonganGaji->jabatan_id = $validated['jabatan_id'];
+        if(trim($request->input("keterangan")) != $potonganGaji->keterangan) {
+            $validated = $request->validate([
+                'keterangan' => 'required|string|unique:'.PotonganGaji::class,
+                'nilai_potongan' => 'required|numeric|gte:0'
+            ]);
+        } else {
+            $validated = $request->validate([
+                'keterangan' => 'required|string',
+                'nilai_potongan' => 'required|numeric|gte:0'
+            ]);
+        }
+        
+        if($potonganGaji->keterangan != 'potongan tidak hadir') {
+            $potonganGaji->keterangan = $validated['keterangan'];
+        }
+
         $potonganGaji->nilai_potongan = $validated['nilai_potongan'];
         $potonganGaji->save();
 
@@ -103,7 +95,11 @@ class PotonganGajiController extends Controller
      */
     public function destroy(PotonganGaji $potonganGaji)
     {
-        $potonganGaji->delete();
+        
+        if($potonganGaji->keterangan != 'potongan tidak hadir') {
+            $potonganGaji->delete();
+        }
+
         return redirect(route('potongan-gaji.index'));
     }
 }
